Finish add new comment, without issues
