+++ src/index.js
@@ -1,5 +1,5 @@
 import './css/index.css';
 import './css/popUp.css';
-import initComment from './modules/comments';
+import { initComment } from './modules/comments';
 
 initComment();
